Support Multiple Docker Registries image push
The task is to support pushing an application build image into multiple docker registries.
To acheive this an api call is created to support new ecr registry creations and save them
to a mongo collection. uri is created internally depending on the aws region.
Two additional fields are added to the Application bo to specify the additional registries
& changes are done accordingly in the create/update application flows.As a part of the task
the changes are also done to move docker related commands to the BuildSpec base class.
Two Boolean variables are introduced which are needed to give support for createBuild flow
and tagBuild/branchBuild decision. Additional registries will be considered to push the image
only during the createBuild flow. In all other flows the default account registry is the only
one considered as a part of the project buildspec.

JiraId: https://capillarytech.atlassian.net/browse/CAP-64193