changes to support stack to artifactory mapping
JavaChanges:
Created a new bo Artifactory to store the registry information. Created
a controller to create/update/getAll Artifactories. Added List of
Artifactory objects field to DeploymentContext object. Added an endpoint
to StackController to update the Stack object, this only updates the single
field artifactories. Currently all artifactories irrespective of the
artifactory of the Artifact object, a new method is added which filters only
those Artifacts whose artifactory is in the list of artifactories of the Stack
object.

See: https://capillarytech.atlassian.net/browse/CAP-64197